requirements.txt
pandas
python-pptx
streamlit
pyarrow==14.0.2
requests==2.31.0
pyairtable
openpyxl==3.1.2
openai==1.12.0
python-dotenv==1.0.1
tqdm==4.66.1